feat: remove unnecessary simpy containers

+++ src/rlsim/engine/utils.py
@@ -27,4 +27,4 @@ def random_number(distribution, params) -> float:
     else:
         value = 0
 
-    return value
+    return np.float32(value)
